backend/requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain
langchain-core
requests
python-dotenv
google-genai 
groq
